Improvements to ashespp from integration with Castor3D.

+++ include/ashespp/Buffer/StagingBuffer.inl
@@ -501,7 +501,7 @@ namespace ashes
 		, UniformBuffer< T > const & buffer
 		, VkPipelineStageFlags dstStageFlags )const
 	{
-		auto elemAlignedSize = buffer.getAlignedSize();
+		auto elemAlignedSize = uint32_t( buffer.getAlignedSize() );
 		doCopyUniformDataToStagingBuffer( data
 			, count
 			, elemAlignedSize );
